Enum list_stacks::Error

#[non_exhaustive]
pub(crate) enum Error {

    AlreadyExistsException(AlreadyExistsException),
    CfnRegistryException(CfnRegistryException),
    ChangeSetNotFoundException(ChangeSetNotFoundException),
    CreatedButModifiedException(CreatedButModifiedException),
    InsufficientCapabilitiesException(InsufficientCapabilitiesException),
    InvalidChangeSetStatusException(InvalidChangeSetStatusException),
    InvalidOperationException(InvalidOperationException),
    InvalidStateTransitionException(InvalidStateTransitionException),
    LimitExceededException(LimitExceededException),
    NameAlreadyExistsException(NameAlreadyExistsException),
    OperationIdAlreadyExistsException(OperationIdAlreadyExistsException),
    OperationInProgressException(OperationInProgressException),
    OperationNotFoundException(OperationNotFoundException),
    OperationStatusCheckFailedException(OperationStatusCheckFailedException),
    StackInstanceNotFoundException(StackInstanceNotFoundException),
    StackNotFoundException(StackNotFoundException),
    StackSetNotEmptyException(StackSetNotEmptyException),
    StackSetNotFoundException(StackSetNotFoundException),
    StaleRequestException(StaleRequestException),
    TokenAlreadyExistsException(TokenAlreadyExistsException),
    TypeConfigurationNotFoundException(TypeConfigurationNotFoundException),
    TypeNotFoundException(TypeNotFoundException),
    Unhandled(Box<dyn Error + Send + Sync + 'static, Global>),

}

All possible error types for this service.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AlreadyExistsException(AlreadyExistsException)

The resource with the name requested already exists.

Tuple Fields of AlreadyExistsException

0: AlreadyExistsException

CfnRegistryException(CfnRegistryException)

An error occurred during a CloudFormation registry operation.

Tuple Fields of CfnRegistryException

0: CfnRegistryException

ChangeSetNotFoundException(ChangeSetNotFoundException)

The specified change set name or ID doesn't exit. To view valid change sets for a stack, use the ListChangeSets action.

Tuple Fields of ChangeSetNotFoundException

0: ChangeSetNotFoundException

CreatedButModifiedException(CreatedButModifiedException)

The specified resource exists, but has been changed.

Tuple Fields of CreatedButModifiedException

0: CreatedButModifiedException

InsufficientCapabilitiesException(InsufficientCapabilitiesException)

The template contains resources with capabilities that weren't specified in the Capabilities parameter.

Tuple Fields of InsufficientCapabilitiesException

0: InsufficientCapabilitiesException

InvalidChangeSetStatusException(InvalidChangeSetStatusException)

The specified change set can't be used to update the stack. For example, the change set status might be CREATE_IN_PROGRESS, or the stack status might be UPDATE_IN_PROGRESS.

Tuple Fields of InvalidChangeSetStatusException

0: InvalidChangeSetStatusException

InvalidOperationException(InvalidOperationException)

The specified operation isn't valid.

Tuple Fields of InvalidOperationException

0: InvalidOperationException

InvalidStateTransitionException(InvalidStateTransitionException)

Error reserved for use by the CloudFormation CLI. CloudFormation does not return this error to users.

Tuple Fields of InvalidStateTransitionException

0: InvalidStateTransitionException

LimitExceededException(LimitExceededException)

The quota for the resource has already been reached.

For information on resource and stack limitations, see Limits in the CloudFormation User Guide.

Tuple Fields of LimitExceededException

0: LimitExceededException

NameAlreadyExistsException(NameAlreadyExistsException)

The specified name is already in use.

Tuple Fields of NameAlreadyExistsException

0: NameAlreadyExistsException

OperationIdAlreadyExistsException(OperationIdAlreadyExistsException)

The specified operation ID already exists.

Tuple Fields of OperationIdAlreadyExistsException

0: OperationIdAlreadyExistsException

OperationInProgressException(OperationInProgressException)

Another operation is currently in progress for this stack set. Only one operation can be performed for a stack set at a given time.

Tuple Fields of OperationInProgressException

0: OperationInProgressException

OperationNotFoundException(OperationNotFoundException)

The specified ID refers to an operation that doesn't exist.

Tuple Fields of OperationNotFoundException

0: OperationNotFoundException

OperationStatusCheckFailedException(OperationStatusCheckFailedException)

Error reserved for use by the CloudFormation CLI. CloudFormation does not return this error to users.

Tuple Fields of OperationStatusCheckFailedException

0: OperationStatusCheckFailedException

StackInstanceNotFoundException(StackInstanceNotFoundException)

The specified stack instance doesn't exist.

Tuple Fields of StackInstanceNotFoundException

0: StackInstanceNotFoundException

StackNotFoundException(StackNotFoundException)

The specified stack ARN doesn’t exist or stack doesn’t exist corresponding to the ARN in input.

Tuple Fields of StackNotFoundException

0: StackNotFoundException

StackSetNotEmptyException(StackSetNotEmptyException)

You can't yet delete this stack set, because it still contains one or more stack instances. Delete all stack instances from the stack set before deleting the stack set.

Tuple Fields of StackSetNotEmptyException

0: StackSetNotEmptyException

StackSetNotFoundException(StackSetNotFoundException)

The specified stack set doesn't exist.

Tuple Fields of StackSetNotFoundException

0: StackSetNotFoundException

StaleRequestException(StaleRequestException)

Another operation has been performed on this stack set since the specified operation was performed.

Tuple Fields of StaleRequestException

0: StaleRequestException

TokenAlreadyExistsException(TokenAlreadyExistsException)

A client request token already exists.

Tuple Fields of TokenAlreadyExistsException

0: TokenAlreadyExistsException

TypeConfigurationNotFoundException(TypeConfigurationNotFoundException)

The specified extension configuration cannot be found.

Tuple Fields of TypeConfigurationNotFoundException

0: TypeConfigurationNotFoundException

TypeNotFoundException(TypeNotFoundException)

The specified extension does not exist in the CloudFormation registry.

Tuple Fields of TypeNotFoundException

0: TypeNotFoundException

Unhandled(Box<dyn Error + Send + Sync + 'static, Global>)

An unhandled error occurred.

Tuple Fields of Unhandled

0: Box<dyn Error + Send + Sync + 'static, Global>